Complete Teensy microcontroller projects exploring interfacing, sampling, and controlling the device
Specific constraints, challenges, and attributes that are unique to embedded systems are explored. The role andcapabilities of such compact systems in today's world is analyzed, with detailed reference to available processingspeed, storage and power consumption. Students gain knowledge of sensors and actuators, their operationalparameters and characteristics. Methods for interfacing, sampling, and controlling the devices are coded, includingpolled vs. interrupt driven, D to A and A to D conversions and simple digital vs. PWM outputs.